+++ tools/issue-labeler/src/SearchIndexCreator/Program.cs
@@ -6,7 +6,7 @@ using Octokit;
 using Azure.Identity;
 using Azure.AI.OpenAI;
 using Azure.Search.Documents.Indexes;
-using AzureRAGService;
+using AzureRagService;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 
@@ -27,7 +27,6 @@ namespace SearchIndexCreator
             Console.WriteLine("2. Process Docs");
             Console.WriteLine("3. Process Issue Examples");
             Console.WriteLine("4. Process Demo");
-            Console.WriteLine("5. Test RAG");
 
             var input = Console.ReadLine();
 
@@ -46,9 +45,6 @@ namespace SearchIndexCreator
                         break;
                     case "4":
                         await ProcessDemo(config);
-                        break;
-                    case "5":
-                        await TestRAG(config);
                         break;
                     default:
                         Console.WriteLine("Invalid option selected.");
@@ -138,82 +134,6 @@ namespace SearchIndexCreator
 
                 await client.Issue.Create("jeo02", "issue-examples", newIssue);
             }
-        }
-
-
-        //Just to mess with my RAG methods in different ways
-        private static async Task TestRAG(IConfigurationSection config)
-        {
-            // Configurations for correct access in search and OpenAI
-            var credential = new DefaultAzureCredential();
-            var searchEndpoint = new Uri(config["SearchEndpoint"]);
-            var openAIEndpoint = new Uri(config["OpenAIEndpoint"]);
-            string modelName = config["OpenAIModelName"];
-
-            // Configuration for Issue specifics
-            string issueIndexName = config["IssueIndexName"];
-            string issueSemanticName = config["IssueSemanticName"];
-            string issueFieldName = "text_vector";
-
-            // Configuration for Document specifics
-            string documentIndexName = config["DocumentIndexName"];
-            string documentSemanticName = config["DocumentSemanticName"];
-            string documentFieldName = "text_vector";
-
-            // Search prompt
-            string searchPrompt = "Azure.AI.OpenAI";
-
-            // Initialize the RAG service
-            var loggerFactory = LoggerFactory.Create(builder => builder.AddConsole());
-            var logger = loggerFactory.CreateLogger<TriageRAG>();
-            var ragService = new TriageRAG(logger);
-
-            // Top X documents/issues
-            int top = 5;
-
-            var relevantIssues = ragService.AzureSearchQuery<AzureRAGService.Issue>(
-                searchEndpoint, issueIndexName, issueSemanticName, issueFieldName, credential, searchPrompt, top
-            );
-
-            var relevantDocuments = ragService.AzureSearchQuery<Document>(
-                searchEndpoint, documentIndexName, documentSemanticName, documentFieldName, credential, searchPrompt, top
-            );
-
-            var docs = relevantDocuments.Select(rd => new
-            {
-                Content = rd.Item1.ToString(),
-                Score = rd.Item2
-            });
-
-            var issues = relevantIssues.Select(ri => new
-            {
-                Content = ri.Item1.ToString(),
-                Score = ri.Item2
-            });
-
-            string docContent = JsonConvert.SerializeObject(docs);
-            string issueContent = JsonConvert.SerializeObject(issues);
-
-            string instructions = "You are an AI assistant designed to generate realistic GitHub issues relevant to the documentation.";
-            string message = $"Documentation: {docContent}\nExample Issues: {issueContent}\nPlease generate potential questions that could be asked about the documentation as GitHub issue queries. Do not provide a response to the questions, only provide a list of detailed questions. Use the example issues on guidance on how to make an issue.";
-
-            // Structure of the response passed to the StructureMessage method.
-            BinaryData structure = BinaryData.FromBytes("""
-                        {
-                          "type": "object",
-                          "properties": {
-                            "Category": { "type": "string" },
-                            "Service": { "type": "string" },
-                            "Response": { "type": "string" }
-                          },
-                          "required": [ "Category", "Service", "Response" ],
-                          "additionalProperties": false
-                        }
-                        """u8.ToArray());
-
-            string result = ragService.SendMessageQna(openAIEndpoint, credential, modelName, instructions, message, structure);
-
-            Console.WriteLine($"Open AI Response:\n{result}");
         }
     }
 }